feature(dao)：分库分表

+++ dts-client/src/main/java/io/github/jasonnull/dts/client/biz/AdminBiz.java
@@ -50,6 +50,6 @@ public interface AdminBiz {
      * @param jobId
      * @return
      */
-    public ReturnT<String> triggerJob(int jobId);
+    public ReturnT<String> triggerJob(Long jobId);
 
 }
